+++ eslint.config.js
@@ -17,7 +17,7 @@ const __dirname = path.dirname(__filename);
 const gitignorePath = path.resolve(__dirname, ".gitignore");
 
 export default defineConfig(
-  globalIgnores(["public/assets/lib/**/*"]),
+  globalIgnores(["src/templates/public/assets/lib/**/*"]),
   includeIgnoreFile(gitignorePath),
   js.configs.recommended,
   ...pluginVue.configs["flat/recommended"],
@@ -45,7 +45,7 @@ export default defineConfig(
     },
   },
   {
-    files: ["public/assets/qrcode.html"],
+    files: ["src/templates/public/assets/qrcode.html"],
     plugins: { html },
     languageOptions: {
       sourceType: "script",