+++ src/templates/scripts/pages/category.ts
@@ -1,0 +1,2 @@
+import "../main-with-header";
+import "/styles/mixins/id-archive.css";