+++ src/templates/scripts/pages/tag.ts
@@ -1,3 +1,3 @@
 import "../main-with-header";
 import "../../styles/pages/tag.css";
-import "/src/styles/mixins/id-archive.css";
+import "/styles/mixins/id-archive.css";